fix(injector): handle synthetic AIDL metadata
Synthetic keystore2 binders did not handle AIDL metadata transactions,
so getInterfaceHash() and getInterfaceVersion() returned unknown
transaction.

Resolve the keystore2 AIDL version from VINTF, fall back to Android
version mapping when needed, and return the matching frozen AIDL hash
from precomputed constants. Keep metadata transaction decoding separate
from business method decoding and reject mismatched interface tokens.

+++ src/main.rs
@@ -139,7 +139,7 @@ fn create_rpc_server() -> Result<Arc<RpcServer>> {
     server.set_authorizer(|peer| {
         let allowed = matches!(
             peer,
-            PeerIdentity::Local { uid, .. } if *uid == 0 || *uid == KEYSTORE_UID as u32
+            PeerIdentity::Local { uid, .. } if *uid == 0 || *uid == KEYSTORE_UID
         );
         if !allowed {
             warn!("Rejected OMK RPC peer {peer}");
